Redis npm commands, test refactors
