+++ package.json
@@ -8,6 +8,7 @@
     "test": "lerna run test",
     "test:debug": "lerna run test:debug",
     "test:integration": "lerna run test:integration",
+    "test:cache": "lerna run test:cache",
     "test:core": "lerna run --scope @commerce-apps/core test:debug",
     "test:ec": "lerna run --scope @commerce-apps/exchange-connector test:debug",
     "test:generator": "lerna run --scope commerce-sdk test:debug",
