+++ packages/core/package.json
@@ -6,7 +6,12 @@
   "types": "dist/index.d.ts",
   "scripts": {
     "pretest": "npm run lint",
+    "redis": "docker rm --force redis-commerce-sdk; docker run -dp 6379:6379 --name redis-commerce-sdk redis:5-alpine",
+    "redis:shutdown": "docker rm --force redis-commerce-sdk; docker rm --force redis-commerce-sdk-debug",
+    "redis:debug": "docker rm --force redis-commerce-sdk-debug; docker run -dp 6379:6379 --name redis-commerce-sdk-debug redis:5-alpine && docker exec -it redis-commerce-sdk-debug redis-cli monitor",
     "test": "nyc mocha \"test/**/*.test.ts\" --reporter=xunit --reporter-options output=../../reports_integration/core.xml",
+    "test:cache": "nyc --check-coverage=false mocha \"test_integration/**/*.ts\" --reporter=xunit --reporter-options output=../../reports/core-integration.xml --retries 2 --timeout 5000",
+    "test:integration": "npm run test:cache",
     "test:debug": "nyc mocha \"test/**/*.test.ts\"",
     "build": "tsc",
     "lint": "eslint \"**/*.ts\" --quiet --fix"
@@ -102,8 +107,10 @@
   "gitHead": "2a6031f5f57c4ba5df8bfb5c4ca42b2631befe58",
   "dependencies": {
     "@commerce-apps/exchange-connector": "^1.3.0-alpha.9",
+    "@keyv/redis": "^2.1.1",
     "dotenv": "~8.2.0",
     "jsonwebtoken": "^8.5.1",
+    "keyv": "^4.0.0",
     "lodash": "~4.17.15",
     "make-fetch-happen": "~7.1.0",
     "node-fetch": "~2.6.0",
